+++ main/src/sdlog.hpp
@@ -1,6 +1,8 @@
 #pragma once
 #include <stdio.h>
 
+#define SDTAG "SD"
+
 // full data structure
 typedef struct data_packet_t{
   int64_t timestamp;
@@ -10,4 +12,5 @@ typedef struct data_packet_t{
 } DataPacket;
 
 void logWriteHeader(FILE** file);
-void logWrite(FILE** file, DataPacket* sdpacket);
+void logWrite(FILE** file, DataPacket* datapacket);
+void taskSD(void *datapacket);
