+++ main/src/telemetry.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+// LoRa packet struct
+typedef struct lora_packet{
+  uint8_t packetid;
+  int baro;
+  float temp;
+} LoraPacket;
+
+void taskLoRa_tx(void *pvParameters);
